Feature/firmware updater (#34)
* Added Firmware updater with versioning in ino files, endpoint to get current firmware version and a way to flash the firmware

* restore versionnumbers

* Fix for #21. Restarting function is not working

* Removed unneeded logic

* Restored Home button

* Added clear sideways pattern and unified clear options

* Remove options file for now

* ; typo fix

* fix version commit

* Integrated Firmware updater with new way of fetching firmware and motor driver versions

* Added proof of concept of Currently Playing

* Moved and improved Currently Playing

* Various styling improvements, moved font locally and added icon font

* Various styling improvements, moved settings to separate overlay.

* Icon fixes

* Software updater functionality and Updated Changelog

* Improved handling and notification of software update

* Disabled debug

* Installing correct required dependencies via DockerFile

* Minor styling fixes and ux improvements

* Using pre-compiled firmware files

* Added esp32 bin

* Used esp32dev board selection in arduino ide

* fix set_speed, allow for changing speed mid pattern

* add changelog

---------
